Structuring folder and functions
