+++ index.html
@@ -54,31 +54,11 @@
           </div>
         </article>
       </section>
-      <!--  -->
-      <!-- <section class="container_article" id="uploadingDog">
-        <h2 class="title title_upload">Upload your Cat or Dog Image</h2>
-        <form class="container-btn" action="" id="uploadingForm">
-          <input class="input_upload" id="file" type="file" name="file" />
-          <button class="btn btn_upload_cat" type="button">Send Image</button>
-        </form>
-      </section> -->
-      <!--  -->
       <h2 class="title_favorites title">Favorites Cats</h2>
       <section
         class="container_article container_favorites grid"
         id="favorite_cats"
-      >
-        <!-- <div class="container container-favorites-pet">
-          <div class="sub-container">
-            <div class="container-img">
-              <img class="img" src="" alt="Image Random" />
-            </div>
-            <div class="container_btn_remove container-btn">
-              <button class="btn_remove btn">Remove to favorite</button>
-            </div>
-          </div>
-        </div> -->
-      </section>
+      ></section>
       <h2 class="title_favorites title">Favorites Dogs</h2>
       <section
         class="container_article container_favorites grid"
@@ -87,6 +67,6 @@
       <!--  -->
     </main>
 
-    <script src="./main.js"></script>
+    <script async type="module" src="./main.js"></script>
   </body>
 </html>
